move AlignmentPipelineParameters to spark.pipeline package in preparation for introduction of steps
