+++ render-ws-spark-client/src/main/java/org/janelia/render/client/spark/MipmapClient.java
@@ -15,11 +15,11 @@ import org.janelia.alignment.spec.stack.StackId;
 import org.janelia.alignment.spec.stack.StackWithZValues;
 import org.janelia.render.client.ClientRunner;
 import org.janelia.render.client.RenderDataClient;
-import org.janelia.render.client.parameter.AlignmentPipelineParameters;
 import org.janelia.render.client.parameter.CommandLineParameters;
 import org.janelia.render.client.parameter.MipmapParameters;
 import org.janelia.render.client.parameter.MultiProjectParameters;
 import org.janelia.render.client.parameter.RenderWebServiceParameters;
+import org.janelia.render.client.spark.pipeline.AlignmentPipelineParameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
